Refactor GitHub to use dedicated class
