+++ .env
@@ -2,8 +2,6 @@ AWS_S3_BUCKET=
 AWS_S3_PREFIX=
 CCAO_REGISTRY_URL=ghcr.io/ccao-data
 GH_APP_ID=
-GH_REPO=
-GH_WORKFLOW=
 IPTS_HOSTNAME=
 IPTS_PORT=
 IPTS_SERVICE_NAME=
