+++ src/utils/spark.py
@@ -6,11 +6,7 @@ import boto3
 from pyarrow import dataset as ds
 from pyspark.sql import SparkSession
 
-from utils.helpers import (
-    create_jwt_token,
-    dispatch_gh_worfklow,
-    strip_table_prefix,
-)
+from utils.helpers import strip_table_prefix
 
 
 class SharedSparkSession:
@@ -31,9 +27,6 @@ class SharedSparkSession:
         database_url: The JDBC URL for the database connection. Constructed
             from the above attributes.
         ipts_password: The password for the database, read from file.
-        gh_app_id: GitHub Application ID for running a workflow.
-        gh_repo: API URL for the target repository containing a workflow.
-        gh_workflow: Workflow YAML file, relative to `.github/workflows`.
         fetch_size: The fetch size for the database queries. This is a tuning
             parameter for query speed. ~10,000 seems to work best.
         initial_compression: The compression type for the initial Parquet files
@@ -46,12 +39,9 @@ class SharedSparkSession:
         spark: The Spark session object.
     """
 
-    def __init__(
-        self, app_name: str, password_file_path: str, gh_pem_path: str
-    ) -> None:
+    def __init__(self, app_name: str, password_file_path: str) -> None:
         self.app_name = app_name
         self.password_file_path = password_file_path
-        self.gh_pem_path = gh_pem_path
 
         # Vars here are loaded from the .env file, then forwarded to the
         # container in docker-compose.yaml
@@ -64,12 +54,6 @@ class SharedSparkSession:
             f"{self.ipts_port}/"
             f"{self.ipts_service_name}"
         )
-
-        # GitHub credentials for launching a target workflow once all jobs
-        # are finished running
-        self.gh_app_id = os.getenv("GH_APP_ID")
-        self.gh_repo = os.getenv("GH_REPO")
-        self.gh_workflow = os.getenv("GH_WORKFLOW")
 
         # Load runtime secret using Compose secrets setup
         with open(self.password_file_path, "r") as file:
@@ -86,21 +70,6 @@ class SharedSparkSession:
         self.s3_prefix = os.getenv("AWS_S3_PREFIX", "iasworld")
 
         self.spark = SparkSession.builder.appName(self.app_name).getOrCreate()
-
-    def run_gh_workflow(self):
-        """
-        Method to run a GitHub Action workflow once all jobs are completed.
-        """
-        if (
-            self.gh_app_id
-            and self.gh_pem_path
-            and self.gh_repo
-            and self.gh_workflow
-        ):
-            jwt = create_jwt_token(self.gh_app_id, self.gh_pem_path)
-            dispatch_gh_worfklow(jwt, self.gh_repo, self.gh_workflow)
-        else:
-            print("GitHub credentials not set, skipping workflow dispatch")
 
 
 class SparkJob:
